- fixed an issue where Java shorts were not handling unsigned short lengths

+++ projects/unitt-websocket/trunk/src/main/java/com/unitt/framework/websocket/WebSocketFragment.java
@@ -283,7 +283,7 @@ public class WebSocketFragment
                         return;
                     }
 
-                    dataLength = new Short( WebSocketUtil.convertBytesToShort( buffer, index ) ).longValue();
+                    dataLength = new Integer( WebSocketUtil.convertBytesToShort( buffer, index ) ).longValue();
                     index += 2;
                 }
                 else if ( dataLength == 127 )
